+++ crates/ruma-client-api/src/delayed_events/update_delayed_event.rs
@@ -3,44 +3,50 @@
 //! Send a delayed event update. This can be a updateing/canceling/sending the associated delayed
 //! event.
 
-pub mod unstable {
+use ruma_common::serde::StringEnum;
+
+use crate::PrivOwnedStr;
+
+/// The possible update actions we can do for updating a delayed event.
+#[derive(Clone, StringEnum)]
+#[ruma_enum(rename_all = "lowercase")]
+#[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
+pub enum UpdateAction {
+    /// Restart the delayed event timeout. (heartbeat ping)
+    Restart,
+    /// Send the delayed event immediately independent of the timeout state. (deletes all
+    /// timers)
+    Send,
+    /// Delete the delayed event and never send it. (deletes all timers)
+    Cancel,
+
+    #[doc(hidden)]
+    _Custom(PrivOwnedStr),
+}
+
+pub mod unstable_v1;
+
+pub mod unstable_v2 {
     //! `msc3814` ([MSC])
     //!
     //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4140
 
     use ruma_common::{
-        api::{auth_scheme::AccessToken, request, response},
+        api::{auth_scheme::NoAccessToken, request, response},
         metadata,
-        serde::StringEnum,
     };
 
-    use crate::PrivOwnedStr;
+    use super::UpdateAction;
 
     metadata! {
         method: POST,
         rate_limited: true,
-        authentication: AccessToken,
+        authentication: NoAccessToken,
         history: {
-            unstable("org.matrix.msc4140") => "/_matrix/client/unstable/org.matrix.msc4140/delayed_events/{delay_id}",
+            unstable("org.matrix.msc4140") => "/_matrix/client/unstable/org.matrix.msc4140/delayed_events/{delay_id}/{action}",
         }
     }
 
-    /// The possible update actions we can do for updating a delayed event.
-    #[derive(Clone, StringEnum)]
-    #[ruma_enum(rename_all = "lowercase")]
-    #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-    pub enum UpdateAction {
-        /// Restart the delayed event timeout. (heartbeat ping)
-        Restart,
-        /// Send the delayed event immediately independent of the timeout state. (deletes all
-        /// timers)
-        Send,
-        /// Delete the delayed event and never send it. (deletes all timers)
-        Cancel,
-
-        #[doc(hidden)]
-        _Custom(PrivOwnedStr),
-    }
     /// Request type for the [`update_delayed_event`](crate::delayed_events::update_delayed_event)
     /// endpoint.
     #[request]
@@ -49,6 +55,7 @@ pub mod unstable {
         #[ruma_api(path)]
         pub delay_id: String,
         /// Which kind of update we want to request for the delayed event.
+        #[ruma_api(path)]
         pub action: UpdateAction,
     }
 
@@ -72,7 +79,7 @@ pub mod unstable {
     }
 
     #[cfg(all(test, feature = "client"))]
-    mod tests {
+    mod client_tests {
         use std::borrow::Cow;
 
         use ruma_common::api::{
@@ -81,6 +88,7 @@ pub mod unstable {
         use serde_json::{Value as JsonValue, json};
 
         use super::{Request, UpdateAction};
+
         #[test]
         fn serialize_update_delayed_event_request() {
             let supported = SupportedVersions {
@@ -91,7 +99,7 @@ pub mod unstable {
                 Request::new("1234".to_owned(), UpdateAction::Cancel)
                     .try_into_http_request(
                         "https://homeserver.tld",
-                        SendAccessToken::IfRequired("auth_tok"),
+                        SendAccessToken::None,
                         Cow::Owned(supported),
                     )
                     .unwrap();
@@ -99,14 +107,43 @@ pub mod unstable {
             let (parts, body) = request.into_parts();
 
             assert_eq!(
-                "https://homeserver.tld/_matrix/client/unstable/org.matrix.msc4140/delayed_events/1234",
+                "https://homeserver.tld/_matrix/client/unstable/org.matrix.msc4140/delayed_events/1234/cancel",
                 parts.uri.to_string()
             );
             assert_eq!("POST", parts.method.to_string());
             assert_eq!(
-                json!({"action": "cancel"}),
+                json!({}),
                 serde_json::from_str::<JsonValue>(std::str::from_utf8(&body).unwrap()).unwrap()
             );
+        }
+    }
+
+    #[cfg(all(test, feature = "server"))]
+    mod server_tests {
+
+        use ruma_common::api::IncomingRequest;
+
+        use super::{Request, UpdateAction};
+
+        #[test]
+        fn deserialize_update_delayed_events_request() {
+            let uri = http::Uri::builder()
+                .scheme("https")
+                .authority("matrix.org")
+                .path_and_query(
+                    "/_matrix/client/unstable/org.matrix.msc4140/delayed_events/a_delay_id/send",
+                )
+                .build()
+                .unwrap();
+
+            let req = Request::try_from_http_request(
+                http::Request::builder().method("POST").uri(uri).body("").unwrap(),
+                &["a_delay_id", "send"],
+            )
+            .unwrap();
+
+            assert_eq!(req.delay_id, "a_delay_id".to_owned());
+            assert_eq!(req.action, UpdateAction::Send);
         }
     }
 }